third_party/blink/web_tests/fast/events/mouse-focus-imagemap-expected.txt

This tests that a link in an image map with tabindex < 0 is not mouse focusable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.activeElement.id is 'focusable-area'
PASS document.activeElement.id == 'focusable-area' || document.activeElement.id == 'body' is true
PASS successfullyParsed is true

TEST COMPLETE
